modules/root_kafka/requirements.txt
PyMuPDF
Flask
Flask-API==3.0.post1
Flask-Cors==3.0.10
kafka-python
nltk
pymongo
pynvml
python-docx
requests
sumy
Werkzeug
python-dotenv==1.0.0
aspose-words
